Precompute report mapping context
